automation/requirements.txt
PyYAML>=6.0
rich>=13.7.0
textual>=0.63.0
